--- packages/app/src/app/pages/Sandbox/Editor/Workspace/Git/Changes/elements.js
@@ -1,26 +1,0 @@
-import styled, { css } from 'styled-components';
-import { EntryContainer } from '../../elements';
-
-export const ChangeContainer = styled.div`
-  &:last-child {
-    border-bottom: none;
-  }
-`;
-
-export const Entry = styled(EntryContainer)`
-  display: flex;
-  align-items: center;
-  line-height: 1;
-  color: ${props => props.theme.vscodeTheme.colors['editor.foreground']};
-  ${({ hideColor }) =>
-    hideColor &&
-    css`
-      background-color: transparent;
-      padding-left: 0;
-    `};
-
-  svg {
-    color: ${({ color }) => color};
-    margin-right: 0.5rem;
-  }
-`;
