+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/GitHub/CreateRepo/index.ts
@@ -1,0 +1,1 @@
+export { CreateRepo } from './CreateRepo';
