+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/GitHub/Git/Changes/index.ts
@@ -1,0 +1,3 @@
+export { Added } from './Added';
+export { Deleted } from './Deleted';
+export { Modified } from './Modified';
