+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/GitHub/Git/TotalChanges/index.ts
@@ -1,0 +1,1 @@
+export { TotalChanges } from './TotalChanges';
