+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/GitHub/Git/index.ts
@@ -1,0 +1,1 @@
+export { Git } from './Git';
